requirements.txt
abstract==2021.11.29.1
asgiref==3.4.1
author==0.0.0.0.dev20200926
base32hex==1.0.2
blank==0.1.0
certifi==2021.10.8
charset-normalizer==2.0.10
choices==0.1
colouration==2020.7.26
content==0.0.1
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.3
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-ninja==0.16.1
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
graphviz==0.19.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
openapi-codec==1.3.2
orjson==3.6.5
packaging==21.3
params==0.9.0
Pillow==9.0.0
pydantic==1.8.2
pyparsing==3.0.7
pytz==2021.3
redis==4.1.1
requests==2.27.1
sche==1.0a5
simplejson==3.17.6
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
